Correct bad-practice in synopsis and clarify usage
State explicitly that the object gets passed into the sub
(like a method call) and show this usage pattern clearly in Synopsis.

Instead of using $self from the closure use the passed-in value
to avoid creating an un-reclaimable reference to the instance
(which won't be garbage collected until global destruction).